+++ src/test/java/selenium/genericshopwebapptests/utils/DriverFactory.java
@@ -1,0 +1,34 @@
+package selenium.genericshopwebapptests.utils;
+
+import io.github.bonigarcia.wdm.WebDriverManager;
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.chrome.ChromeDriver;
+import org.openqa.selenium.edge.EdgeDriver;
+import org.openqa.selenium.firefox.FirefoxDriver;
+
+public class DriverFactory {
+
+    private static WebDriver driver;
+
+    public static WebDriver getDriver() {
+        driver = selectDriver("Chrome");
+        return driver;
+    }
+
+    public static WebDriver selectDriver(String browserName) {
+
+        if (browserName.equalsIgnoreCase("Chrome")) {
+            WebDriverManager.chromedriver().setup();
+            return new ChromeDriver();
+        } else if (browserName.equalsIgnoreCase("Firefox")) {
+            WebDriverManager.firefoxdriver().setup();
+            return new FirefoxDriver();
+        } else if (browserName.equalsIgnoreCase("Edge")) {
+            WebDriverManager.edgedriver().setup();
+            return new EdgeDriver();
+        } else {
+            System.out.println("Incorrect browser name");
+        }
+        return null;
+    }
+}
