+++ src/test/java/selenium/seleniumcodingexercises/utils/DriverType.java
@@ -1,0 +1,8 @@
+package selenium.seleniumcodingexercises.utils;
+
+public enum DriverType {
+
+    CHROME,
+    FIREFOX,
+    EDGE
+}
