--- src/test/java/seleniumcodingexercises/utils/DriverType.java
@@ -1,8 +1,0 @@
-package seleniumcodingexercises.utils;
-
-public enum DriverType {
-
-    CHROME,
-    FIREFOX,
-    EDGE
-}
